// front-end/src/api/models/agendamentoControllerHorariosAdministradorParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API Agendador de Barbearia
 * Contrato OpenAPI da API REST do Agendador de Barbearia.
 * OpenAPI spec version: 1.0
 */

export type AgendamentoControllerHorariosAdministradorParams = {
/**
 * @minimum 1
 */
idCliente: number;
/**
 * @minimum 1
 */
idBarbeiro: number;
/**
 * @minimum 1
 */
idFilial: number;
/**
 * @pattern ^\d{4}-\d{2}-\d{2}$
 */
data: string;
/**
 * Array JSON de itens {idServico,quantidade,desconto}.
 * @minItems 1
 */
servicos: string;
};