// front-end/src/api/models/horariosDisponiveisRespostaDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API Agendador de Barbearia
 * Contrato OpenAPI da API REST do Agendador de Barbearia.
 * OpenAPI spec version: 1.0
 */
import type { HorarioDisponivelRespostaDto } from './horarioDisponivelRespostaDto';

export interface HorariosDisponiveisRespostaDto {
  fuso: string;
  duracaoTotalMinutos: number;
  valorTotal: string;
  horarios: HorarioDisponivelRespostaDto[];
}